Pass allow_off parameter to DecreaseScreenBrightness dbus method
Modify the dbus_method_call1 to allow specifying the type of the
parameter.  Used to be able to pass int to the user_activity
method and bool to the DecreaseScreenBrightness method.   The
DecreaseScreenBrightness parameter allow_off is always set to
false, as it doesn't make much sense to allow the screen to go
completely off while frecon is active.

BUG=chromium:441558
TEST=test brightness up/down on link while frecon is active

Change-Id: Iced5e36ff819643f5f4f74d1199a1d996aee0a03
Reviewed-on: https://chromium-review.googlesource.com/258151
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: David Sodman <dsodman@chromium.org>
Trybot-Ready: David Sodman <dsodman@chromium.org>
Tested-by: David Sodman <dsodman@chromium.org>