Frecon: bugfixes for UDL.
We should only consider drivers with at least one connected connector in
kms_open().

If we didn't find a connector with the desired type, return the first one
that is connected. If none are connected, give up and return to avoid a
NULL pointer reference.

Force drmModeDirtyFB to damage the rect so that screen is refreshed
on UDL.

BUG=chromium:475165
TEST=frecon on UDL on link

Change-Id: I0860b12de7bec1afc660395cd47bd0230c5bc84d
Reviewed-on: https://chromium-review.googlesource.com/264795
Tested-by: Haixia Shi <hshi@chromium.org>
Commit-Queue: Haixia Shi <hshi@chromium.org>
Trybot-Ready: Haixia Shi <hshi@chromium.org>
Reviewed-by: David Sodman <dsodman@chromium.org>