Add terminal properties to splash screen
In order to support text display and input on the splash screen for
powerwash and dev/normal transitions, the splash screen needs to be
attached to a terminal.  This change creates a terminal for the splash
screen and directs the name of the pts device to stdout so that
the scripts that launch the terminal can access the terminal.

BUG=chromium:450400
TEST=test on link with looping splash screen

Change-Id: I9ab746f67627a8258bf87920f3c462ef4341ff10
Reviewed-on: https://chromium-review.googlesource.com/269211
Reviewed-by: Haixia Shi <hshi@chromium.org>
Tested-by: David Sodman <dsodman@chromium.org>
Commit-Queue: David Sodman <dsodman@chromium.org>