frecon: read result of Chrome DBUS calls and act on them.
When Chrome takes ownership of display, it runs display reconfiguration.
During that time it will ignore and fail calls to release ownership of display.
Frecon did not read result of DBUS function calls and assumed they succeeded
when they did not. This change fixes that and makes frecon retry the calls if
they fail.

BUG=chromium:658280
TEST=switch to VT2 then switch back to Chrome and quickly back to VT2.

Change-Id: I3335ed3dcea260a6da69675b713e4546c6e4d4d0
Signed-off-by: Dominik Behr <dbehr@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/403615
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>