frecon: Fix documentation on loop-count
The loop-count argument specifies setting it to zero causes
it to loop forever.  Setting zero, in fact, causes it to not
play back at all.  Rather than change the behavior of the
program by fixing this behavior, change the documentation to
match the behavior.

BUG=chromium:770346
TEST=No functional change.

Change-Id: Id47e7e83622fc1e01dcc32288955ccc80aeb9cfd
Reviewed-on: https://chromium-review.googlesource.com/693047
Commit-Ready: Pat Erley <perley@chromium.org>
Tested-by: Dominik Behr <dbehr@chromium.org>
Tested-by: Pat Erley <perley@chromium.org>
Reviewed-by: Dominik Behr <dbehr@chromium.org>